balena-fin-overlay: Update for kernel 5.4
The base for these overlays file is the upstream linux-raspberrypi
kernel overlay at tag `raspberrypi-kernel_1.20201022-1`:
https://github.com/raspberrypi/linux/blob/raspberrypi-kernel_1.20201022-1/arch/arm/boot/dts/overlays/balena-fin-overlay.dts

It then includes the following modifications:

* Use pwrseq-simple to power on the SD8787 WiFi.

In meta-balena we use a modified built-in pwrseq-sd8777 driver. The
advantage of the pwrseq-simple driver is that it is already part
of the Raspbian kernel. The difference between them is that the
pwrseq-sd8777 driver introduces a 300ms delay between powering
on the chip and the PDN line assertion done in the post_power_on
hook. This strictly complies with the Marvell chip
specification.

https://github.com/balena-os/balena-raspberrypi/blob/v2.60.1%2Brev6/layers/meta-balena-raspberrypi/recipes-kernel/linux/linux-raspberrypi/0004-mmc-pwrseq-Repurpose-for-Marvell-SD8777.patch

* Switch balenaFin WiFi to STA+uAP mode (uap0 interface)
* Add DTS overrides to config.txt:

  `wifi_dbg` could be used for tuning the verbosity of the WiFi driver.

  `wifi_mode` could be used for activating/deactivating WiFi interfaces. For
   example `dtparam=wifi_mode=1` will only expose the `wlan0` interface.

The upstream patch in meta-balena is:
https://github.com/balena-os/balena-raspberrypi/blob/v2.60.1+rev6/layers/meta-balena-raspberrypi/recipes-kernel/linux/linux-raspberrypi/0005-balena-fin-wifi-sta-uap-mode.patch

For more information refer to the user manual:
https://github.com/balena-io/sd8887-mrvl/tree/master/wlan_src

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>