Add simple npm test to satisfy Flowzone
Technically this is an npm repo (it builds the PDFs) and Flowzone
noticed. When it tried to run the npm tests, later versions of npm fail
if there's no test specified. This is an attempt to satisfy the
requirement for a test, using the simplest, representative test possible
- just run the program. If it runs successfully, the test passes.

Change-type: patch
Signed-off-by: Heath Raftery (@hraftery) <heath.raftery@balena.io>